+++ activerecord/test/cases/counter_cache_test.rb
@@ -65,6 +65,15 @@ class CounterCacheTest < ActiveRecord::TestCase
     end
   end
 
+  test "reset counters for multiple records" do
+    t1, t2 = topics(:first, :second)
+    Topic.increment_counter(:replies_count, [t1.id, t2.id])
+
+    assert_difference ["t1.reload.replies_count", "t2.reload.replies_count"], -1 do
+      Topic.reset_counters([t1.id, t2.id], :replies_count)
+    end
+  end
+
   test "reset multiple counters" do
     Topic.update_counters @topic.id, replies_count: 1, unique_replies_count: 1
     assert_difference ["@topic.reload.replies_count", "@topic.reload.unique_replies_count"], -1 do
@@ -165,6 +174,16 @@ class CounterCacheTest < ActiveRecord::TestCase
 
     assert_difference "subscriber.reload.books_count", -1 do
       Subscriber.reset_counters(subscriber.id, "books")
+    end
+  end
+
+  test "reset counter of polymorphic association" do
+    aircraft = Aircraft.create!
+    aircraft.wheels << Wheel.create!
+    Aircraft.increment_counter(:wheels_count, aircraft.id)
+
+    assert_difference "aircraft.reload.wheels_count", -1 do
+      Aircraft.reset_counters(aircraft.id, :wheels_count)
     end
   end